Split tests that require circular dependencies
https://github.com/jsimonetti/rtnetlink and
https://github.com/mdlayher/ethtool depend on netlink and therefore
introduce a circular dependency. This library was packaged for Debian
and circular dependencies should be avoided there (e.g. to ease
bootstrapping).

`rtnetlink` and `ethtool` are only needed for integration testing. So
move all tests that require `rtnetlink` or `ethtool` into
`conn_linux_consumer_test.go` and allow skipping running those test by
specifying the build tag `break_circular_dependencies`.
